API updates
- Change LangValue to an i16
- Remove temporary userId field from a few GQL fields
- Change output of machine state to make it more friendly to the UI
- Change websocket output from snake_case to camelCase

+++ api/migrations/2020-03-11-204852_init_models/up.sql
@@ -10,8 +10,8 @@ CREATE TABLE program_specs (
     id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
     slug VARCHAR(20) NOT NULL,
     hardware_spec_id UUID NOT NULL REFERENCES hardware_specs(id),
-    input INTEGER[] NOT NULL,
-    expected_output INTEGER[] NOT NULL,
+    input SMALLINT[] NOT NULL,
+    expected_output SMALLINT[] NOT NULL,
     UNIQUE(slug, hardware_spec_id)
 );
